ai-web/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
PyMySQL==1.1.0
python-dotenv==1.0.0
bcrypt==4.1.2
Werkzeug==3.0.1
WTForms==3.1.1
email-validator==2.1.0
requests>=2.31.0
numpy>=2.0.0
opencv-python>=4.12.0
gunicorn
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
pytest-mock==3.12.0
faker==20.1.0
pytest-xdist==3.5.0
pytest-timeout==2.2.0